fix(auth): Review-Runde 1 — Löschsperre über alle RESTRICT-Referenzen, offene Links bei Sperre/Passwort/Adresswechsel geschlossen, SPACE_DELETED für den persönlichen Space
Weiter: Seitenarithmetik ohne int-Überlauf und Obergrenze für page, Sortierrichtung und Suchlänge
per Allowlist, Ablauf-Erinnerung am Kalendertag in der Zeitzone des Laufs und nur für aktive
Konten, eine Verwalter-Wiedervorlage je Lauf (AdminReviewReminderStep vereint Ablauf- und
Quartalszählung), Batch-Laden und Obergrenze je Lauf in den Tagesschritten, Race bei gleichzeitiger
Anlage derselben Adresse als 409, neutraler EMAIL_TAKEN-Text, unlock nur bei tatsächlicher Sperre,
ACCOUNT_UNLOCKED ohne hängenden Link-Satz, resetCaches wieder package-private (MailTestSupport),
redeem liefert die verbrauchte Zeile, Testaufräumen der MAIL_SETTINGS_CHANGED-Zeilen; Spec und
Feature-Doku nachgezogen (Löschen praktisch nur für nie benutzte Konten, PATCH-Ablauf in der
Vergangenheit, Admin-Endpunkte profilunabhängig).

Refs #1537

+++ backend/src/main/java/io/opaa/auth/local/LocalUserAdminService.java
@@ -26,6 +26,12 @@ import org.springframework.stereotype.Service;
  * link whose delivery event is missing - the request answers 500, and the administrator's repeat
  * supersedes the link and writes the event. Deliberately not {@code @Transactional}: a send of up
  * to the SMTP timeouts must never hold a database connection.
+ *
+ * <p>The list is filtered, sorted and paged in memory over every local account of the organization:
+ * the state is derived ({@code LocalCredentials#state}) and the activity a class, both one rule
+ * that a second SQL formulation would only duplicate, and the population is small by design
+ * (ADR-0033: local accounts are the exception, a page holds at most 50). A repository query with
+ * {@code Pageable} is the step to take once an installation carries thousands of them.
  */
 @Service
 public class LocalUserAdminService {
@@ -46,8 +52,8 @@ public class LocalUserAdminService {
             .filter(overview -> matches(overview, query))
             .sorted(comparator(query))
             .toList();
-    int from = Math.min(query.page() * query.size(), matching.size());
-    int to = Math.min(from + query.size(), matching.size());
+    int from = (int) Math.min((long) query.page() * query.size(), matching.size());
+    int to = (int) Math.min((long) from + query.size(), matching.size());
     return new LocalUserPage(
         matching.subList(from, to), matching.size(), query.page(), query.size());
   }
